enable compilation
