--- include/operator/cpu/cpu_operators.hpp
@@ -1,7 +1,0 @@
-#ifndef CPU_OPERATORS_HPP
-#define CPU_OPERATORS_HPP
-
-#include "operator/operator.hpp"
-
-
-#endif // CPU_OPERATORS_HPP
