+++ src/operator/cpu/sub_operator_impl.cpp
@@ -1,5 +1,5 @@
 #include "operator/operators.hpp"
-#include "operator/aux_operator/elementwise_operator.hpp"
+#include "operator/cpu/elementwise_operator.hpp"
 
 namespace CPU_OP
 {
