+++ src/operator/div_operator.cpp
@@ -1,6 +1,6 @@
 #include <functional>
 #include "operator/operators.hpp"
-#include "operator/aux_operator/elementwise_operator.hpp"
+#include "operator/cpu/elementwise_operator.hpp"
 
 std::vector<std::vector<size_t>> DivOperator::inferOutputShapes(const std::vector<Tensor> &inputs,
                                                                 const std::unordered_map<std::string, Node::AttributeValue> &attributes)
